refactor: view 패키지 구조 기능별로 분리
