+++ app/src/androidTest/java/woowacourse/movie/BookingSummaryActivityTest.kt
@@ -14,8 +14,8 @@ import org.junit.jupiter.api.DisplayName
 import org.junit.runner.RunWith
 import woowacourse.movie.model.DefaultPricingPolicy
 import woowacourse.movie.model.MovieTicket
-import woowacourse.movie.ui.view.BookingActivity
-import woowacourse.movie.ui.view.BookingSummaryActivity
+import woowacourse.movie.ui.view.booking.BookingActivity
+import woowacourse.movie.ui.view.booking.BookingSummaryActivity
 import java.time.LocalDateTime
 
 @RunWith(AndroidJUnit4::class)
