+++ app/src/main/AndroidManifest.xml
@@ -14,13 +14,13 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".ui.view.BookingActivity"
+            android:name=".ui.view.booking.BookingActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.view.BookingSummaryActivity"
+            android:name=".ui.view.booking.BookingSummaryActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.view.MainActivity"
+            android:name=".ui.view.main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
