+++ app/src/main/java/woowacourse/movie/ui/view/booking/BookingSummaryActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui.view
+package woowacourse.movie.ui.view.booking
 
 import android.os.Bundle
 import android.widget.TextView
@@ -8,9 +8,10 @@ import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import woowacourse.movie.R
 import woowacourse.movie.model.MovieTicket
-import woowacourse.movie.ui.view.CustomFormatter.formatAmount
-import woowacourse.movie.ui.view.CustomFormatter.formatDateTime
-import woowacourse.movie.ui.view.CustomFormatter.formatHeadCount
+import woowacourse.movie.ui.view.utils.CustomFormatter.formatAmount
+import woowacourse.movie.ui.view.utils.CustomFormatter.formatDateTime
+import woowacourse.movie.ui.view.utils.CustomFormatter.formatHeadCount
+import woowacourse.movie.ui.view.utils.intentSerializable
 
 class BookingSummaryActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -30,7 +31,8 @@ class BookingSummaryActivity : AppCompatActivity() {
     }
 
     private fun displayBookingSummary() {
-        val movieTicket = intent.intentSerializable(getString(R.string.ticket_info_key), MovieTicket::class.java)
+        val movieTicket =
+            intent.intentSerializable(getString(R.string.ticket_info_key), MovieTicket::class.java)
         val title = findViewById<TextView>(R.id.title)
         val screeningDateTime = findViewById<TextView>(R.id.screeningDateTime)
         val headCount = findViewById<TextView>(R.id.headCount)
@@ -38,7 +40,8 @@ class BookingSummaryActivity : AppCompatActivity() {
 
         title.text = movieTicket.title
         screeningDateTime.text = formatDateTime(movieTicket.screeningDateTime)
-        headCount.text = formatHeadCount(getString(R.string.headCount_message), movieTicket.headCount)
+        headCount.text =
+            formatHeadCount(getString(R.string.headCount_message), movieTicket.headCount)
         amount.text = formatAmount(getString(R.string.amount_message), movieTicket.amount)
     }
 }
