+++ app/src/main/res/layout/activity_bookingsummary.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.view.BookingSummaryActivity">
+    tools:context=".ui.view.booking.BookingSummaryActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
